feat: implement "Disable tray" option
Add a "Disable Tray" checkbox, with its state saved and loaded to
the data store. The state of the tray icon is managed by this
checkbox at runtime. Depending on the visibility of the tray icon,
the application will either minimize to the tray or close completely
upon the close event.